refactor(webapp): align create-organization modal with shared dialog design
Rebuild the create-organization modal on the same shell as the create-project
and create-experiment dialogs: ModalHeader, FormField, PrimaryButton and the
bordered 560px panel, replacing the shadcn DialogHeader/Separator/Label/Input
layout with its outline Cancel button.

Submission now goes through a real form with toast.promise, and the button is
disabled while loading or while the name is blank, so the manual "name is
required" toast is no longer needed. Also drops a duplicated "Organization
Description" label block and renames the file to kebab-case to match the other
components.
